v0.01
okay after actually reading the lab, we change how fast the reverb is
modulated by having it run in a separate ISR like we did on the lab
quiz. Methods are fine for what I'm doing, Just need to push it into a
second interrupt thats frequency is controlled by the keypad.